+++ packages/packages/core/editor-interactions/src/components/controls/direction.tsx
@@ -20,7 +20,7 @@ export function Direction( { value, onChange }: FieldProps ) {
 					{ __( 'Direction', 'elementor' ) }
 				</Typography>
 			</Grid>
-			<Grid item xs={ 12 } md={ 6 }>
+			<Grid item xs={ 12 } md={ 6 } sx={ { display: 'flex', justifyContent: 'flex-end', overflow: 'hidden' } }>
 				<ToggleButtonGroup
 					size="tiny"
 					exclusive
